+++ client/src/elements/schema-editor/schema-editor-wrapper.js
@@ -1,9 +1,14 @@
 import { bindable, inject } from "aurelia-framework";
+import { LogManager } from "aurelia-framework";
 import { getType } from "./helpers";
+import mixinDeep from "mixin-deep";
 import NotificationChecker from "resources/checkers/NotificationChecker.js";
 import AvailabilityChecker from "resources/checkers/AvailabilityChecker.js";
+import ToolEndpointChecker from "resources/checkers/ToolEndpointChecker.js";
 
-@inject(NotificationChecker, AvailabilityChecker, Element)
+const log = LogManager.getLogger("Q");
+
+@inject(NotificationChecker, AvailabilityChecker, ToolEndpointChecker, Element)
 export class SchemaEditorWrapper {
   @bindable
   schema;
@@ -22,9 +27,15 @@ export class SchemaEditorWrapper {
 
   options = {};
 
-  constructor(notificationChecker, availabilityChecker, element) {
+  constructor(
+    notificationChecker,
+    availabilityChecker,
+    toolEndpointChecker,
+    element
+  ) {
     this.notificationChecker = notificationChecker;
     this.availabilityChecker = availabilityChecker;
+    this.toolEndpointChecker = toolEndpointChecker;
     this.element = element;
     this.getType = getType;
   }
@@ -33,8 +44,7 @@ export class SchemaEditorWrapper {
     // Clear visible notification if previous notification was of type Required
     if (
       this.visibleNotification &&
-      this.visibleNotification.message.title ===
-      "notifications.required.title"
+      this.visibleNotification.message.title === "notifications.required.title"
     ) {
       this.visibleNotification = "";
     }
@@ -44,8 +54,24 @@ export class SchemaEditorWrapper {
     if (this.schema.hasOwnProperty("Q:options")) {
       this.options = Object.assign(this.options, this.schema["Q:options"]);
     }
+    // dynamicSchema is only handled once, it's not possible to overwrite the dynamicSchema using dynamicSchema :-)
+    if (this.options.dynamicSchema && !this.dynamicSchemaApplied) {
+      if (this.options.dynamicSchema.type !== "ToolEndpoint") {
+        throw new Error(
+          `${
+            this.options.dynamicSchema.type
+          } is not implemented as dynamicSchema type`
+        );
+      }
+      this.dynamicSchemaApplied = true;
+      this.applyDynamicSchema();
+      this.reevaluateDynamicSchemaCallback = this.applyDynamicSchema.bind(this);
+      this.toolEndpointChecker.registerReevaluateCallback(
+        this.reevaluateDynamicSchemaCallback
+      );
+    }
 
-    // if this has notificationChecks
+    // handle notificationChecks
     if (Array.isArray(this.options.notificationChecks)) {
       this.applyNotifications();
       this.reevaluateNotificationsCallback = this.applyNotifications.bind(this);
@@ -54,7 +80,7 @@ export class SchemaEditorWrapper {
       );
     }
 
-    // if this has availabilityChecks
+    // handle availabilityChecks
     if (Array.isArray(this.options.availabilityChecks)) {
       this.applyAvailability();
       this.reevaluateAvailabilityCallback = this.applyAvailability.bind(this);
@@ -125,6 +151,31 @@ export class SchemaEditorWrapper {
           this.element
         );
       }
+    }
+  }
+
+  async getDynamicSchema(dynamicSchema) {
+    try {
+      return await this.toolEndpointChecker.check(dynamicSchema.config);
+    } catch (e) {
+      throw new Error(
+        `failed to get dynamicSchema for ${JSON.stringify(dynamicSchema)} ${e}`
+      );
+    }
+  }
+
+  async applyDynamicSchema() {
+    try {
+      const dynamicSchema = await this.getDynamicSchema(
+        this.options.dynamicSchema
+      );
+      // make a copy of the schema so the schemaChangedCallbacks get applied
+      this.schema = Object.assign({}, mixinDeep(this.schema, dynamicSchema));
+    } catch (e) {
+      log.error(
+        `Failed to assign dynamicSchema to schema, you need to figure out why this happened as this case is not handled in a nice way and could feel pretty weird.`,
+        e
+      );
     }
   }
 }