+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 4.0.0 (26.06.2019)
+
+- breaking: The support for the deprecated config style for AvailabilityChecks and NotificationChecks is removed. All config needs to follow the new style according to the docs in README
+- feature: `dynamicSchema` is a new property in `Q:options` in the JSON schema for schema-editor. It allows to dynamically alter the schema based on data. See README for details.
+- feature: schema-editor-object schemas support `expandable: true` in `Q:options` resulting in the object property inputs being hidden at first, expandable by a button (as is already supported for array entries)
+- feature: `Q:options.buttonLabel` can be used on array properties to define the label used on the add/delete buttons
+- feature: visual hierarchy is improved in the options pane
+- feature: improve the login error messages in case the server request times out and remove the message if the browser is not Chrome (since all evergreen browsers are supported by now)
+- deprecated: `dynamicEnum` is deprecated and will be removed in v5. Use the new `dynamicSchema` to implement the same functionality.
+
 # 3.8.2 (22.01.2019)
 
 - fix: preview container doesn't grow in height across it's container anymore